All CRUD operations added.

+++ src/main/java/demo/springboot_API_project_1/controllers/CustomerController.java
@@ -28,13 +28,14 @@ public class CustomerController {
 		if (customers != null)
 			return new ResponseEntity<List<Customer>>(customers, HttpStatus.ACCEPTED);
 		else {
-			return new ResponseEntity<String>("no records found", HttpStatus.NO_CONTENT);
+			return new ResponseEntity<String>("no records found", HttpStatus.OK);
 		}
 	}
 
 	@PostMapping("/customer")
 	public ResponseEntity<?> addCustomer(@RequestBody Customer customer) {
 		try {
+			repository.save(customer);
 			return new ResponseEntity<String>("User successfully registered", HttpStatus.CREATED);
 		} catch (Exception ex) {
 			return ResponseEntity.badRequest().body("invalid key or value..\n" + ex.getMessage());
@@ -42,29 +43,42 @@ public class CustomerController {
 	}
 
 	@GetMapping("/customer/{id}")
-	public ResponseEntity<?> get_user(@PathVariable int id) {
+	public ResponseEntity<?> getCustomer(@PathVariable int id) {
 		Optional<Customer> customer = repository.findById(id);
 		if (customer.isPresent()) {
 			return ResponseEntity.ok(customer.get());
 		}
-		return ResponseEntity.badRequest().body("Customer not found");
+		return new ResponseEntity<String>("Customer not found", HttpStatus.BAD_REQUEST);
 	}
 
 	@PutMapping("/customer/{id}")
-	public Customer update_user(@PathVariable int id, @RequestBody Customer customer) {
+	public ResponseEntity<?> updateCustomer(@PathVariable int id, @RequestBody Customer customer) {
 		Optional<Customer> _customer = repository.findById(id);
 		if (_customer.isPresent()) {
-			Customer customer_db = _customer.get();
-			customer_db.update_customer(customer);
-			return repository.save(customer_db);
+			try {
+				Customer customer_db = _customer.get();
+				customer_db.update_customer(customer);
+				repository.save(customer_db);
+				return new ResponseEntity<String>("details updated successfully", HttpStatus.ACCEPTED);
+			} catch (Exception ex) {
+				return ResponseEntity.badRequest().body("invalid key or value..\n" + ex.getMessage());
+			}
 		}
-		return null;
+		return new ResponseEntity<String>("Customer record not found", HttpStatus.BAD_REQUEST);
 	}
 
 	@DeleteMapping("/customer/{id}")
-	public String delete_user(@PathVariable int id) {
-		repository.deleteById(id);
-		return "User deleted!!";
+	public ResponseEntity<?> deleteCustomer(@PathVariable int id) {
+		Optional<Customer> _customer = repository.findById(id);
+		if (_customer.isPresent()) {
+			try {
+				repository.deleteById(id);
+				return new ResponseEntity<String>("User deleted!!", HttpStatus.ACCEPTED);
+			} catch (Exception ex) {
+				return ResponseEntity.badRequest().body("invalid key or value..\n" + ex.getMessage());
+			}
+		}
+		return new ResponseEntity<String>("Customer record not found", HttpStatus.BAD_REQUEST);
 	}
 
 }
